initial 4

+++ src/main/java/ua/nure/hrynko/SummaryTask4/db/dto/Roles.java
@@ -4,26 +4,22 @@ import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
+
 @Entity(name = "roles")
 public class Roles {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @NotNull
     @Size(max = 256)
     @Column(name = "name", nullable = false)
     private String name;
 
-    public void setId(Long id) { this.id = id; }
-    public Long getId() {
-        return id;
-    }
+    public void setId(int id) {this.id = id;}
+    public int getId() {return id;}
 
-    public void setName(String name) {
-        this.name = name;
-    }
+    public void setName(String name) {this.name = name;}
     public String getName() {
         return name;
     }
